feat(*) moar improvements
* refactor(computed) move sections and size to computed props. It didn't
make much sense to have the computedSections being calculated manually
so now it is a computed prop. This helps with perf since computed props
have memoization. This also allows to also remove computedSize so we can
remove side effects and difficult to understand computations. This also
completely removes the watcher, which is something that has needed to
happen since I introduced it.
* fix(initialValue) since we removed initialValue, make v-model work.
* refactor(getSize) rename getSize to computed prop totalResults
* docs(readme) update readme to have all the new stuff v-model, etc
* docs(storybook) update stories to match new feats